fix(2_0_1_7): widen ACAuthResponsePacket account id to 8 bytes
The 2.0.1.7 Trion client decodes ACAuthResponsePacket as:

  ulong accountId    (8 bytes)
  byte[32] wsk       (32 bytes)
  byte slotCount     (1 byte)

Vanilla writes accountId.Value where AccountId.Value is uint (4 bytes).
With the old layout the client reads 8 bytes for accountId, which
consumes 4 real bytes + 4 bytes of wsk, shifting every following field
by 4 bytes. slotCount lands inside wsk, the trailing 4 bytes overflow
the packet, and the client throws a 'Serializer Mismatch' popup right
after Connect.

Cast the value to ulong before writing on this branch:

  stream.Write((ulong)accountId.Value);

Matches the old 2.0.1.7 alpha source's ACAuthResponsePacket (which
declared _accountId as ulong directly). The on-the-wire width changes
from 4 to 8 bytes; this is intentional and a 1.2 client cannot read
the new layout, but the 1.2 branch on master is unaffected.

+++ AAEmu.Login/Core/Packets/L2C/ACAuthResponsePacket.cs
@@ -15,7 +15,11 @@ public class ACAuthResponsePacket(AccountId accountId, byte slotCount) : LoginPa
 
     public override PacketStream Write(PacketStream stream)
     {
-        stream.Write(accountId.Value);
+        // 2.0.1.7 expects an 8-byte ulong account id. AccountId.Value is uint
+        // (4 bytes) on the master branch — writing it raw on this branch
+        // shifts wsk + slotCount by 4 bytes and the client throws a
+        // "Serializer Mismatch" popup right after Connect. Cast explicitly.
+        stream.Write((ulong)accountId.Value);
         stream.Write(_wsk, true);
         stream.Write(slotCount);
